Add slice click listener implementation and refactor code in cleanup
Implemented the handling of click events for the PieChart in the sample project and refactored the code to rename classes
and add comments etc
