+++ opensrp-reporting/src/main/java/org/smartregister/reporting/interfaces/IndicatorVisualisationFactory.java
@@ -3,7 +3,7 @@ package org.smartregister.reporting.interfaces;
 import android.content.Context;
 import android.view.View;
 
-import org.smartregister.reporting.models.ReportingIndicatorVisualization;
+import org.smartregister.reporting.model.ReportingIndicatorVisualization;
 
 public interface IndicatorVisualisationFactory {
     View getIndicatorView(ReportingIndicatorVisualization data, Context context);
